split component into smaller components
