+++ app/dashboard/dashboard-components/project-details/project-info.tsx
@@ -1,0 +1,15 @@
+import React from "react";
+import { ProjectResponseDto } from "../../types/types";
+
+interface ProjectInfoProps {
+  project: ProjectResponseDto;
+}
+
+export function ProjectInfo({ project }: ProjectInfoProps) {
+  return (
+    <div>
+      <h4 className="text-sm font-semibold">Project ID</h4>
+      <p className="text-sm text-muted-foreground">{project.id}</p>
+    </div>
+  );
+}
